# orchestrator/api.yaml
openapi: 3.1.0
info:
  title: Orchestrator API
  description: |
    HTTP API сервиса оркестрации игровых серверов.

    Оркестратор управляет вычислительными нодами (game-server-node) через gRPC,
    хранит метаданные в PostgreSQL и предоставляет REST API для:
    - управления серверными билдами,
    - запуска и остановки инстансов игровых серверов,
    - просмотра журналов в реальном времени,
    - обнаружения серверов клиентами игр (discovery).
  version: 1.0.0
servers:
  - url: /api/orchestrator/v1
    description: Базовый путь оркестратора
...
paths:
...
  /nodes:
    post:
      summary: Подключить вычислительную ноду
      description: |
        Разработчик подключает собственный вычислительный узел к платформе.

        Два сценария:
        1. **Ручной ввод IP** — разработчик знает адрес ноды и передаёт его
           вместе с авторизационным токеном. Оркестратор проверяет доступность
           ноды по gRPC и добавляет в реестр.
        2. **Авторизация обнаруженной ноды** — нода уже вещает и видна в списке
           со статусом `unauthorized` (неавторизована). Разработчик передаёт
           только `node_id` и токен — оркестратор авторизует ноду.
      operationId: registerNode
      tags:
        - nodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - title: manual
                  description: Ручное добавление ноды по адресу
                  type: object
                  required:
                    - address
                    - token
                  properties:
                    address:
                      type: string
                      description: gRPC-адрес ноды (host:port)
                      example: "192.168.1.100:50051"
                    token:
                      type: string
                      description: |
                        Авторизационный токен ноды. Устанавливается при
                        запуске агента на ноде (флаг `--node-token` или
                        переменная окружения `NODE_TOKEN`).
                    region:
                      type: string
                      description: Регион ноды (опционально)
                  additionalProperties: false
                - title: authorize
                  description: Авторизация обнаруженной ноды
                  type: object
                  required:
                    - node_id
                    - token
                  properties:
                    node_id:
                      type: integer
                      format: int64
                      description: ID ноды, обнаруженной в реестре
                    token:
                      type: string
                      description: Токен для авторизации ноды
                  additionalProperties: false
      responses:
        "201":
          description: Нода подключена и авторизована
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Неверный токен авторизации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Нода с указанным ID не найдена в реестре
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Нода уже авторизована или уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: Список всех нод
      description: Возвращает все зарегистрированные вычислительные ноды.
      operationId: listNodes
      tags:
        - nodes
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/NodeStatus"
          description: Фильтр по статусу ноды
      responses:
        "200":
          description: Список нод
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Node"
        "500":
          $ref: "#/components/responses/InternalError"
  /nodes/{nodeId}:
    get:
      summary: Информация о ноде
      operationId: getNode
      tags:
        - nodes
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Информация о ноде
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Удалить ноду
      description: |
        Удаляет ноду из реестра. Все инстансы на этой ноде переводятся
        в статус crashed.
      operationId: deleteNode
      tags:
        - nodes
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Нода удалена
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /nodes/{nodeId}/usage:
    get:
      summary: Потребление ресурсов ноды
      operationId: getNodeUsage
      tags:
        - nodes
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Метрики ноды
          content:
            application/json:
              schema:
                type: object
                properties:
                  node_id:
                    type: integer
                    format: int64
                  usage:
                    $ref: "#/components/schemas/ResourceUsage"
                  active_instance_count:
                    type: integer
                    format: int32
                    description: Количество активных инстансов (starting, running, stopping)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ═══════════════════════════════════════════
  #  HEALTH
  # ═══════════════════════════════════════════
...
components:
  schemas:
...
    NodeStatus:
      type: string
      enum:
        - unauthorized
        - online
        - offline
        - maintenance
      description: |
        Статус вычислительной ноды.
        `unauthorized` — нода обнаружена, но ещё не авторизована разработчиком.

    # ─── Билды ────────────────────────────────
...
    Node:
      type: object
      properties:
        id:
          type: integer
          format: int64
        address:
          type: string
          description: gRPC-адрес ноды (host:port)
        region:
          type: string
        status:
          $ref: "#/components/schemas/NodeStatus"
        cpu_cores:
          type: integer
          format: int32
        total_memory_bytes:
          type: integer
          format: int64
        total_disk_bytes:
          type: integer
          format: int64
        agent_version:
          type: string
          description: Версия агента на ноде
        last_ping_at:
          type: string
          format: date-time
          description: Время последнего heartbeat
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ─── Ресурсы ──────────────────────────────
    ResourceUsage:
      type: object
      properties:
        cpu_usage_percent:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: Загрузка CPU в процентах
        memory_used_bytes:
          type: integer
          format: int64
        disk_used_bytes:
          type: integer
          format: int64
        network_bytes_per_sec:
          type: integer
          format: int64
          description: Суммарный входящий и исходящий трафик
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Машинно-читаемый код ошибки
          example: "BUILD_IN_USE"
        message:
          type: string
          description: Человекочитаемое описание
          example: "Cannot delete build: 2 running instances depend on it"
        details:
          type: object
          additionalProperties: true
          description: Дополнительные детали (опционально)
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"